requirements.txt
asgiref==3.2.10
Django==2.0.6
django-filter==1.1.0
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
future==0.18.2
mysqlclient==2.0.1
Pillow==7.2.0
PyJWT==1.7.1
python-http-client==3.2.7
python-slugify==4.0.0
pytz==2020.1
sendgrid==6.3.1
sqlparse==0.3.1
text-unidecode==1.3
